(refactor)[home-manager/scripts]: Improve commit handling logic
Improved the commit handling logic in the rebuildSystem script to only run the 'git commit' command if there are changes to be committed. If there are no changes, the script now outputs a message indicating that there are no commits to make. This change ensures that unnecessary commits are not created when there are no changes in the git repository.