(refactor)[home-manager/scripts]: Refactor tmuxSessionizer.nix
Added a new shell script to manage tmux sessions in NixOS, allowing users to easily switch between different projects using tmux. The script checks if a tmux session is already running and creates a new session based on the selected project directory. If a session with the same name already exists, it switches to that session instead.